Move enum Status definition to types.h
