+++ src/bus/uart/uart.h
@@ -4,6 +4,8 @@
 #include <stdbool.h>
 #include <stdint.h>
 
+#include "types.h"
+
 #include "../../commands.h"
 
 /**
