+++ src/registers/converters/ctmu.c
@@ -1,3 +1,5 @@
+#include "types.h"
+
 #include "../../commands.h"
 #include "../../helpers/delay.h"
 #include "ctmu.h"
