+++ src/registers/converters/ctmu.h
@@ -3,6 +3,8 @@
 
 #include <xc.h>
 
+#include "types.h"
+
 #ifdef	__cplusplus
 extern "C" {
 #endif /* __cplusplus */
