+++ src/sdcard/sdcard.h
@@ -3,6 +3,8 @@
 
 #include <stdint.h>
 
+#include "types.h"
+
 #include "../commands.h"
 
  /**
