+++ src/transport/packet_handler.h
@@ -1,6 +1,7 @@
 #ifndef PSLAB_PACKET_HANDLER_H
 #define PSLAB_PACKET_HANDLER_H
 
+#include "types.h"
 #include "../commands.h"
 
 enum Status PACKET_exchange(void);
