+++ src/transport/uart2.h
@@ -1,5 +1,6 @@
 #include <stdint.h>
 
+#include "types.h"
 #include "../commands.h"
 
 enum Status UART2_read(
